Preserve register %a2.
If we don't preserver the %a2 register through MacOS traps, the content
of "emile_l2_header_t* info" in second/main:start() is lost.

It seems gcc supposes this register is not modified whereas it is.

This has been detected on a powermac 7100. I don't know if this appears
because of the powermac ROM or of the gcc version (4.3.2).

Signed-off-by: Laurent Vivier <laurent@vivier.eu>